+++ populate_db.py
@@ -6,7 +6,24 @@ import datetime as dt
 from models import *
 import init_db as idb
 
-fake=Faker("fr_FR")
+fake = Faker('fr_FR')
+
+# Fonction pour créer des membres fictifs
+def creation_membres(nombre_de_membres: int):
+    membres = []
+    for _ in range(nombre_de_membres):
+        nom = fake.name()
+        email = fake.email()
+        membre = Membre(nom_membre=nom, email=email)
+        membres.append(membre)
+    return membres
+
+# Générer des membres fictifs et les enregistrer dans la base de données
+def enregistrer_membres(nombre_de_membres: int, engine: Engine):
+    membres = creation_membres(nombre_de_membres)
+    with sm.Session(engine) as session:
+        session.add_all(membres)
+        session.commit()
 
 
 #creation des faux membres(pour 10)
